Some fixes for deployment... still not working. Needs more testing. See notes.org for more features, too.

+++ setup.py
@@ -15,11 +15,13 @@ parser.add_argument('-p', '--pack', action='store_true',
                     default=False, help='Pack app files into archive.')
 parser.add_argument('-b', '--build', action='store_true',
                     default=False, help='Build executable.')
+parser.add_argument('-m', '--make', action='store_true',
+                    default=False, help='Make fresh database file.')
 
 def clean():
-    if os.path.exists('./food_pantry.zip'):
-        print 'Removing archive: ./food_pantry.zip ...'
-        os.remove('./food_pantry.zip')
+    if os.path.exists('./emmaus_house_food_pantry.zip'):
+        print 'Removing archive: ./emmaus_house_food_pantry.zip ...'
+        os.remove('./emmaus_house_food_pantry.zip')
 
     if os.path.exists('./build'):
         print 'Removing directory: ./build ...'
@@ -29,7 +31,14 @@ def clean():
         print 'Removing directory: ./dist ...'
         shutil.rmtree('./dist')
 
-def pack():
+    if os.path.exists('./django.log'):
+        print 'Removing log: ./django.log ...'
+        os.remove('./django.log')
+
+def make():
+    import settings
+    if settings.STANDALONE:
+        raise Exception('STANDALONE setting True while making database')
     if os.path.exists('./pantry.db'):
         print 'Removing previous database: ./pantry.db ...'
         os.remove('./pantry.db')
@@ -38,6 +47,10 @@ def pack():
     print 'Loading data in new database ...'
     subprocess.check_call('python.exe manage.py loaddata data\\initial_v2.json')
 
+def pack():
+    if not os.path.exists('./pantry.db'):
+        raise Exception('Missing database file: ./pantry.db')
+
     print 'Compiling python files ...'
     compileall.compile_dir('.', force=True)
 
@@ -45,16 +58,21 @@ def pack():
     files = []
     for dirpath, dirnames, filenames in os.walk("."):
         for filename in filenames:
+            if dirpath.startswith('.\\.git'): continue
             files.append(os.path.join(dirpath, filename))
 
     print 'Creating archive file ...'
-    with ZipFile('food_pantry.zip', 'w') as food_pantry_zip:
+    with ZipFile('emmaus_house_food_pantry.zip', 'w') as food_pantry_zip:
         for filename in files:
             food_pantry_zip.write(filename)
 
 def build():
+    import settings
+    if not settings.STANDALONE:
+        raise Exception('STANDALONE setting False while building webapp')
+
     print 'Reading archive into memory ...'
-    with open('./food_pantry.zip', 'rb') as food_pantry_zip:
+    with open('./emmaus_house_food_pantry.zip', 'rb') as food_pantry_zip:
         food_pantry_data = food_pantry_zip.read()
 
     print 'Invoking py2exe to generate the executable ...'
@@ -81,6 +99,7 @@ if __name__ == '__main__':
     if args.clean: clean()
     if args.pack: pack()
     if args.build: build()
+    if args.make: make()
 else:
     class DummyArgs:
         def __getattr__(self, field):
